+++ modules/module-mysql/test/src/BinlogStreamUtils.ts
@@ -113,7 +113,6 @@ export class BinlogStreamTestContext {
 
   async replicateSnapshot() {
     await this.binlogStream.initReplication();
-    await this.storage!.autoActivate();
     this.replicationDone = true;
   }