+++ modules/module-postgres/src/replication/WalStream.ts
@@ -321,7 +321,7 @@ export class WalStream {
 
     // Check that replication slot exists
     for (let i = 120; i >= 0; i--) {
-      await touch();
+      this.touch();
 
       if (i == 0) {
         container.reporter.captureException(last_error, {
@@ -479,7 +479,7 @@ WHERE  oid = $1::regclass`,
 
         for (let table of tablesWithStatus) {
           await this.snapshotTableInTx(batch, db, table);
-          await touch();
+          this.touch();
         }
 
         // Always commit the initial snapshot at zero.
@@ -628,7 +628,7 @@ WHERE  oid = $1::regclass`,
         at += rows.length;
         this.metrics.getCounter(ReplicationMetric.ROWS_REPLICATED).add(rows.length);
 
-        await touch();
+        this.touch();
       }
 
       // Important: flush before marking progress
@@ -737,6 +737,9 @@ WHERE  oid = $1::regclass`,
           rows.map((r) => r.key)
         );
       }
+      // Even with resnapshot, we need to wait until we get a new consistent checkpoint
+      // after the snapshot, so we need to send a keepalive message.
+      await sendKeepAlive(db);
     } finally {
       await db.end();
     }
@@ -837,7 +840,6 @@ WHERE  oid = $1::regclass`,
 
   async streamChanges(replicationConnection: pgwire.PgConnection) {
     // When changing any logic here, check /docs/wal-lsns.md.
-
     const { createEmptyCheckpoints } = await this.ensureStorageCompatibility();
 
     const replicationOptions: Record<string, string> = {
@@ -866,9 +868,6 @@ WHERE  oid = $1::regclass`,
     });
 
     this.startedStreaming = true;
-
-    // Auto-activate as soon as initial replication is done
-    await this.storage.autoActivate();
 
     let resnapshot: { table: storage.SourceTable; key: PrimaryKeyValue }[] = [];
 
@@ -911,7 +910,7 @@ WHERE  oid = $1::regclass`,
         let count = 0;
 
         for await (const chunk of replicationStream.pgoutputDecode()) {
-          await touch();
+          this.touch();
 
           if (this.abort_signal.aborted) {
             break;
@@ -1091,11 +1090,10 @@ WHERE  oid = $1::regclass`,
     }
     return Date.now() - this.oldestUncommittedChange.getTime();
   }
-}
 
-async function touch() {
-  // FIXME: The hosted Kubernetes probe does not actually check the timestamp on this.
-  // FIXME: We need a timeout of around 5+ minutes in Kubernetes if we do start checking the timestamp,
-  // or reduce PING_INTERVAL here.
-  return container.probes.touch();
+  private touch() {
+    container.probes.touch().catch((e) => {
+      this.logger.error(`Error touching probe`, e);
+    });
+  }
 }