+++ packages/service-core/src/storage/BucketStorageBatch.ts
@@ -60,17 +60,28 @@ export interface BucketStorageBatch extends ObserverClient<BucketBatchStorageLis
   keepalive(lsn: string): Promise<boolean>;
 
   /**
-   * Set the LSN for a snapshot, before starting replication.
+   * Set the LSN that replication should resume from.
+   *
+   * This can be used for:
+   * 1. Setting the LSN for a snapshot, before starting replication.
+   * 2. Setting the LSN to resume from after a replication restart, without advancing the checkpoint LSN via a commit.
    *
    * Not required if the source database keeps track of this, for example with
    * PostgreSQL logical replication slots.
    */
-  setSnapshotLsn(lsn: string): Promise<void>;
+  setResumeLsn(lsn: string): Promise<void>;
 
   /**
    * Get the last checkpoint LSN, from either commit or keepalive.
    */
   lastCheckpointLsn: string | null;
+
+  /**
+   * LSN to resume from.
+   *
+   * Not relevant for streams where the source keeps track of replication progress, such as Postgres.
+   */
+  resumeFromLsn: string | null;
 
   markSnapshotDone(tables: SourceTable[], no_checkpoint_before_lsn: string): Promise<SourceTable[]>;